Update colours, consolodate search header into search field
