+++ design/src/main/kotlin/uk/govuk/app/design/ui/theme/Color.kt
@@ -109,7 +109,7 @@ data class GovUkColourScheme(
         val buttonCompactHighlight: Color,
         val buttonCompactDisabled: Color,
         val buttonCompactFocused: Color,
-        val searchBox: Color,
+        val search: Color,
         val switchOn: Color,
         val switchOff: Color,
         val toggleHandle: Color,
@@ -184,7 +184,7 @@ internal val LightColorScheme = GovUkColourScheme(
         buttonCompactHighlight = BlueLighter95,
         buttonCompactDisabled = Grey100,
         buttonCompactFocused = YellowPrimary,
-        searchBox = Grey60,
+        search = White,
         switchOn = GreenPrimary,
         switchOff = BlackLighter50,
         toggleHandle = White,
@@ -258,7 +258,7 @@ internal val DarkColorScheme = GovUkColourScheme(
         buttonCompactHighlight = BlueDarker80,
         buttonCompactDisabled = Grey400,
         buttonCompactFocused = YellowPrimary,
-        searchBox = Grey700,
+        search = Black,
         switchOn = GreenPrimary,
         switchOff = BlackLighter50,
         toggleHandle = White,
@@ -333,7 +333,7 @@ internal val LocalColourScheme = staticCompositionLocalOf {
             buttonCompactHighlight = Color.Unspecified,
             buttonCompactDisabled = Color.Unspecified,
             buttonCompactFocused = Color.Unspecified,
-            searchBox = Color.Unspecified,
+            search = Color.Unspecified,
             switchOn = Color.Unspecified,
             switchOff = Color.Unspecified,
             toggleHandle = Color.Unspecified,
